onLoad
в отдельном файле
